BitGo local proxy and signing server
Summary:
The BitGo Express REST API is a lightweight service for developers that want to take advantage of BitGo but are developing in a language without a native BitGo SDK.
BitGo Express should be run as a service on the API consumers’s network (you need to run it, not BitGo).
BitGo Express will proxy typical BitGo REST APIs and implement crypto operations such as key creation, transaction signing and sending coins. These operations can only be performed within the network of the API consumer as they would otherwise expose keys belonging to the user.

Reviewers: mike, ben

Reviewed By: ben

Subscribers: ben

Differential Revision: http://phabricator.bitgo.com/D496