Fix resharing and webhooks tests and API
Summary:
Fixed two issues:

1) The webhooks removeWebhook API call returns the number removed, not the
walletId or other properties.

2) The API has been updated recently to support "resharing", which is when you
need to share the same wallet with someone again because they have lost or
reset their password. To do this, you must set "reshare=true" when sharing the
wallet with them. Older API calls that implicitly relied on resharing now must
be updated to set reshare=true, or they will fail. This commit fixes some tests
that are actually resharing, and thus need to set reshare=true. Also, a check
on the type of the reshare property is added to the shareWallet method.

Reviewers: benchan

Reviewed By: benchan

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D801