seperate create/sign transactions in sdk
Summary:
This is a BREAKING change for exisitng clients that directly interface with createTransaction.
However, most of our clients use sendCoins and not createTransaction so we should be ok (mostly).

The work splits up the createTransaction call and refactors the transaction builder (from class into methods). It also refactors the wallet methods.
In the past, createTransaction both created as well as half-signed the transaction. Now it will only create the transaction, and you need to call signTransaction to sign it.

Reviewers: ryan, ben

Reviewed By: ben

Subscribers: mike, ben

Differential Revision: https://phabricator.bitgo.com/D901