support signing/sending from a safe wallet
Summary:
To add legacy safe wallet support, we need to be able to sign a transaction
with a normal private key rather than with a keychain. The simplest way to do
this was to allow passing in a WIF private key in the keychain variable for
safe wallets (i.e., if wallet.type() === 'safe').

Reviewers: ben, benchan

Reviewed By: benchan

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D1055