Add change splitting functionality to transaction builder
Summary:
Some wallet operators have a problem when operating a hot wallet only
for outgoing transactions because the number of unspents never increases,
meaning if they only have a couple unspents, they end up with long chains
of dependent transactions.

This change introduces the concept of change splitting, which is enabled by
default, but can be explicitly disabled (pass splitChangeSize=0)

An explicit splitChangeSize will split the change as long as the change for
the tx is greater than 2 * splitChangeSize (unless splitChangeSize = 0)

If no value for splitChangeSize is passed, then we query the wallet stats and
use a value which is at the 75th percentile of tx net spend sizes for the wallet.

Reviewers: benchan, jameson

Reviewed By: jameson

Subscribers: barath, huang, alex, mason, mike, ben

Differential Revision: https://phabricator.bitgo.com/D1393