Fix bug with multiple inputs for single key WIF
Summary:
There was a bug in using the single key addresses for paying fees when the inputs were too small. This caused transactions to not be able to be created if you had inputs too small and were using the single key address WIF to pay fees.
This fix shoudl resolve the issue.

Reviewers: ben, alex

Reviewed By: alex

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D2687