Make it possible to pass keychain object to sendMany
Summary:
We want to use sendMany when handling sendmany in BitGoD,
because the travel rule support is bundled there.

In order to use sendMany from BitGoD, we need to be able
to pass the keychain arg directly. We could pass just the xprv
but that will result in an additional unnecessary keychain fetch.

Reviewers: benchan

Reviewed By: benchan

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D2879