fix bug when creating CPFP inputs argument
Summary:
We were passing an array of unspents objects to the server, when it was expecting an array of strings of the form "txHash:outputIndex". This resulted in two different behaviors, depend
ing on if the webclient was making the call, or if the call was being made by the SDK running Node.js.

If the client was the webclient browser, then the "inputs" argument would be converted to a string "[object Object]" which would pass the check on the server and then be ignored

If the client was the SDK running in Node.js, then the array of unspent objects would be converted (on the server) to an Object, which returns the `invalid inputs, expecting string or array of txid strings` error.

Now we extract the "txHash:outputIndex" strings from the unconfirmed unspents, and pass that to getEstimate

Reviewers: abhishek, jameson, arik, benchan

Reviewed By: benchan

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D5145