Prevent IE to cache requests/responses
Summary:
https://bitgoinc.atlassian.net/browse/BG-914
There is an issue with IE where you login/logout and try to login again and in the OTP screen throws an error, this is due to IE caching the requests/responses.
We need this special If-Modified-Since attribute in the header with a very old date so IE don't to that

This was causing the sdk to throw invalid hmac, because it was using an older HMAC for a new request, and thus throwing the error in the client.

Reviewers: arik

Reviewed By: arik

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D6035