clean up unspents method mess by making one method rely on the other and centralize parameter validation
Summary:
Signed-off-by: Arik Sosman <arik@bitgo.com>

https://bitgoinc.atlassian.net/browse/BG-1366

We have two methods for retrieving unspents in V1, wallet.unspents() and wallet.unspentsPaged(). Both end up making the same server method call, and have to duplicate a bunch of the validation logic. This method makes one go through the other, thereby allowing to centralize the validation logic in only one of the two methods.

Reviewers: kevin, alex

Reviewed By: alex

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D7183