BG-3183: Make the invalid environment error friendlier
Summary:
When the environment is invalid, list the supported environments so that
the user doesn't wonder what the options are.

When we log the default environment, say "defaulting to test" instead of
testnet. That way, the user doesn't think that he should call
BitGo.BitGo({ env: testnet }).

Check params and $BITGO_ENV in one place. Some of the validation logic
used to only check params.env, which allowed an invalid environment to
be passed in via $BITGO_ENV.

Reviewers: bigly, zacleids

Reviewed By: zacleids

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D7706