BG-3213: fix bitgo.js tests
Summary:
Clean up tests in test/bitgo.js, so that they pass.  Remove the
try/catch hack in authenticateTestUser().  It was causing failures in
the OTP auth test that expected a 401.

This is an incremental fix to the BitGoJS tests.  There are still
failures in other test suites.  I diffed my output against the tip of
master and didn't see an increase in the number of failed tests.

Reviewers: kevin, zacleids, bigly, tyler

Reviewed By: tyler

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D7739