Add SDK method to recover unsupported ERC20 tokens
Summary:
Add a method which allows a user to recover unspported ERC20 tokens from a wallet.

It requires specifying the token's contract address.

Outputs a 'half-signed' JSON - this is what users will eventually submit to support.

This was tested on testnet with an ERC20 token I created - if you want some to test this  out, I can send it.

Reviewers: tyler, bigly, barath, arik

Reviewed By: barath, arik

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D8021