Add tokens to default constants
Summary:
BG-3700

First of all, let me state for the record that this is a bad and incomplete fix, and does nothing to address the underlying issue of a race condition in getConstants(). This fix defines the missing token fields in the SDK, so that they can be referenced before we get the client constants response from the server. This means that if this list of tokens changes, an associated change in the SDK must be made and rolled out so that those tokens will be instantiated correctly when the SDK initializes.

Fixing the underlying issue would require SDK interface changes, which would necessitate changes in any code that depends on the BitGo SDK (aka code written by our customers). This is a non-starter at this time.

I'm open to other alternative solutions which don't involve requiring dependent code to change.

Signed-off-by: Tyler Levine <tyler@bitgo.com>

Reviewers: bigly, arik

Reviewed By: arik

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D8064