[BG-5293] Add signed flag to recoverFromWrongChain
Summary: Currently, recoverFromWrongChain takes a wallet passphrase or private key to sign the recovery transaction. This is not compatible with our customers who use cold wallets and the offline signing tool. The signed flag defaults to true, but if cleared, creates an unsigned transaction to be signed by the offline signing tool before being sent to support for cosigning.

Reviewers: arik, bigly, kevin

Reviewed By: kevin

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D8878