[BG-5569] Fix bug in recovery process involving LTC canonical addresses
Summary:
When the recovery tool is searching for recoverable outputs on a faulty LTC transaction, it attempts to canonicalize the addresses on each output. In the event that it is unable to convert an address, it exits the loop.

However, there are instances where the first output address is not a valid address (for reasons unknown), but the second output has a valid address which funds can be recovered from.

EX: See this transaction: https://live.blockcypher.com/ltc/tx/b6ab6e2df7761f545fa9942f1511e74332746ec50beb6cf96b51f42f995ba89b/ The first output has no valid address, though the second output can be recovered from.

Changing the break to a continue allows the recovery tool to ignore invalid addresses, rather than fail the recovery process entirely.

Reviewers: arik, bigly, kevin

Reviewed By: kevin

Subscribers: ben

Differential Revision: https://phabricator.bitgo.com/D8993