[BG-5650] Fix/refactor cross-chain recoveries for BCH/LTC
Summary: The recoverFromWrongChain methods in abstractUtxoCoin, bch, and ltc are near-identical, and only vary by an array that specifies which recoveries are allowed on that coin. Refactored this array to be returned from a separate method, so that only one recover method on abstractUtxoCoin is needed, eliminating code reuse and fixing an issue with cold wallet recoveries.

Reviewers: arik, kevin, bigly, mark

Reviewed By: mark

Differential Revision: https://phabricator.bitgo.com/D9028