[BG-6127] Support recoveries when the backup key is held by KRS
Summary:
Allows recovering funds when the user only holds the public key of the backup keypair, and the private key is held by a KRS provider.

When the backupKey is set to an xpub, unspents are gathered as normal, and a transaction to recover the user’s funds is built. The transaction is half-signed with the userKey (which must always be an encrypted xprv). The half-signed transaction hex is exported.

For UTXO coins, a fee specified in config.js is applied. Currently, Keyternal charges a flat $99 USD fee for all recoveries. This amount is converted to crypto using the current market price at bitcoinaverage.com, and added as an output on the transaction. This process fails gracefully if the wallet does not have enough to pay the recovery fee.

For ETH and XRP, no recovery fee output is added. The customer will have to coordinate with the KRS provider to pay before cosigning.

After the half-signed transaction is ready, the public key of the backup keypair is exported. This is provided to the provider so that they may look up the customer’s account information and find the private key.

Reviewers: kevin, bigly, tyler, mtoda

Reviewed By: tyler, mtoda

Subscribers: james, tyler

Differential Revision: https://phabricator.bitgo.com/D9398