BG-8712: Enable OFC coin wallet creation in SDK
Summary:
Quick Summary:
The Offchain (OFC) coin is going to be our coin for wallets were we keep track of their balances ourselves, they don't live on a blockchain.
The server changes are already on master. For creating such wallets, only one key is required to be set (the user key).
This key will sign all transactions and signatures will be verified in our internal accounting system.

Reviewers: bigly, zacleids, kevin, davidcruz, tyler

Reviewed By: tyler

Subscribers: tyler

Differential Revision: https://phabricator.bitgo.com/D10403