Deprecate conflicting params in generateAddress
JIRA: BG-11323

Currently there are 4(!) params in generateAddress that conflict in
specifying what kind of address to create. Chain is the ultimate
source of truth in determining what kind of unspent (and thus what
kind of address) is created by our platform. This commit deprecates
the addressType, segwit, and bech32 params from the generateAddress
function. However, since they are already in the SDK API they are
not removed in this commit, nor is their functionality changed.
There are only two functionality changes to the function. The
first is a check that the addressType derived from addressType,
segwit, and bech32 matches the chain passed. Otherwise an error is
thrown. The second is to check that wrapped segwit is supported
when creating an address for a wrapped segwit unspent. This check
already existed for native segwit and also should have been in
place for wrapped segwit.

We bring in the new @bitgo/unspents library in this commit as well
as it is used to do type checking in generateAddress.
AbstractUtxoCoin.AddressTypes and
AbstractUtxoCoin.AddressTypeChains are elided for canonical enums
defined in this library that will be used across SDK and other
internal bitgo infrastructure.