Update recover() to handle p2wsh
JIRA: BG-11568

Note that the tx ids and the signatures in the recovery tests
changed. This is because this PR also replaces the use of config
constants in fee estimation in favor of the @bitgo/unspents lib.
The worst case estimations are used to be conservative when
calculating the fee. This resulted in output amounts changing,
which in turn changed both the signature and the tx id. This commit
also refactors code used in both `recover()` and
`generateAddress()` into the function `createMultiSigAddress()`.