Replace config constants with unspents lib
JIRA: BG-6452

Note that the config constants were out of sync with the unspents
lib size estimations. Thus, due to the size changes listed below,
tx hexes in tests changed because fee estimations (and
consequently output amounts) changed.

Input Type               | Old Size | New Size
----------------------------------------------
p2sh input               | 296      | 297
p2shP2wsh input          | 139      | 140
uncompressed p2pkh input | 160      | 180

There are a couple of other things to note in regards to fees.
Output size estimation defaults to a p2pkh output, which is NOT the
worst case output size. Non-segwit tx overhead size is also used,
which is also NOT the worst case overhead size. For transactions
that use segwit, and specifically p2wsh outputs, we will be
underestimating fees.

The final thing to note is that there are several places where
chain codes are compared to chains 10 and 11 to determine if the
output is segwit. `isP2shP2wsh()` from the unspents lib is
intentionally not used because it throws if passed an invalid chain
code, which would change functionality and is apparently something
that may happen for v1 safe (non-hd) wallets.