Supply `recipients` in `accelerateTransaction()`
The SDK docs don't mention that the platform API requires an empty
recipients parameter. The unit tests don't capture this because we are
not testing against the real backend.

This fix checks that `recipients` is either undefined or an empty
array (for backward compatibility), and sets it to an empty array.

Issue: BG-12074