Separate signing and verification
JIRA: CT-128

Prior to this commit we could not spend native segwit inputs unless
they were the last input in the transaction. This bug was caused by
the fact that we would rebuild the transaction (and subsequently
clear the redeem script in the p2wsh case) after signing each
input. Then we would verify the input and it would pass for p2wsh
because the redeem script had been cleared. However, when we
rebuilt on the next input, the redeem script would be there again.
We don't reverify the previous input so it slips past verification
and we end up with an invalid tx hex. This commit splits up signing
and signature verification such that this cannot happen.

Side note: I am not sure why bitgo-utxo-lib returns a redeem script
for native segwit inputs. This seems like a bug but the SDK was
already handling this (albeit incorrectly) so maybe it was
expected?