CT-173: Forward all valid parameters to platform for transfers listing
In #337, it was reported that the `type` field was not being propertly
forwarded to the platform server by the `wallet.transfers()` function.

After further investigation, there were two more fields which were also
not properly forwarded.

This change fixes `wallet.transfers()` so that it correctly forwards all
valid parameters to the platform, and updates an existing test so that
we ensure that all valid parameters continue to be passed in the future.