Handle keyreg txes in algo explainTransaction
JIRA: CT-684

This commit checks if the `to` field exists on a tx before
attempting to derive the address from its public key. Prior to this
commit there was no check and `explainTransaction` would blindly
try to derive an address froma transaction's `to` field, which does
not exist on keyreg txes.