Add GitHub Issue Templates
Add two GitHub issues templates to this project (https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository):

1. Bugs: Add a standardized template for contributors to report bugs to
the project maintainers.
2. Feature Requests: Add a template that allows contributors to request
features.

....

Motivation: There is no standardized way for people to report their
issues to this project. When bugs are reported, they tend to omit
critical details such as node version, os, etc. which are extremely
helpful in diagnosing the issue.