Properly deserialize EOS staking transactions
JIRA: ETSTK-11

This commit adds functionality to deserialize EOS staking
transactions. EOS staking transactions consist of two actions,
delegatebw and voteproducer. The delegatebw action specifies the
amount of EOS to be staked and the account to stake it for. In our
case the staker and the receiver of the stake will always be the
same. The voteproducer action specifies the proxy or producers that
the staked funds will vote for.