fix(SERV-593): Correctly handle undefined boolean config items
Boolean flags for bitgo-express (eg, `disableSSL`, `disableProxy`, etc)
currently have no effect.

The issue is that the environment config, which is highest in
precedence, casts undefined env vars to booleans if the config field is
of boolean type. This means the lower precedence configs are not
considered (args and default configs). Therefore, all command line flags
which set booleans are not able to be set, and instead must be set via
environment instead.

fixes #599