Lock transactions to the next block
The locktime denominates the height at which a transaction gets valid
and may be relayed. Only a block with a height greater than the locktime
may include the locktimed transaction.

Consequently, the `nLockTime` should be set to the _current_ height to
allow relaying immediately and to make the transaction eligible to be
included in the next block. For the locktime to be respected, the
`nSequence` on at least one input must be set to `UINT_MAX - 1`. If
`nSequence` on all inputs is set to the maximum value of `0xffffffff`
this signals that a transaction is final and the locktime will be
disregarded. Setting it lower signals BIP-125 replaceability, which we
don't want either at this time.

References:
* https://btcinformation.org/en/developer-guide#locktime-and-sequence-number
* https://bitcoin.stackexchange.com/a/92416/5406
* https://github.com/achow101/btcinformation.org/pull/32

Issue: BLOCK-262