refactor(test): use mergeFixtures
Instead of explicitly merging base fixtures and fork-coin fixtures, we
want a general method that merges them. This reduces the risk of
accidentally forgetting to merge fields and spares us the work from
writing manual merge functions for all fixtures.

The merge rules for merging a baseFixture and a forkFixture are as
follows:

* For every property of forkFixture:
  * If forkFixture property is an object and baseFixture has a
    property with the same name that is also an object, merge both
    fixtures recursively.
  * If forkFixture property is an array and baseFixture has a property
    with the same name that is also an array, append elements to
    baseFixture.

Issue: BG-16466