Fix incorrect type check on string literal
Before our typescript refactor last year, these properties were checked
with the deprecated `common.validateParams`, which takes an array string
literals of the properties to check for on the given object. When we
refactored however, we continued using the string literal instead of the
variable of the same name.

This is the only location in the code where the string `isString('`
appears, so I don't think we have made this mistake anywhere else.

Ticket: BG-18946