Add ENTERPRISE_PAYS_FEES and accountExplorerUrl
A common use case in the front end is building a URL to a block explorer
for a specific transaction, so a user can see the status of their
  transaction on chain.

For some account coins, fees are paid from the Enterprise, instead of
from the wallet itself. For those coins, we'd like to be able to:
1) Know that this is one of those coins where the enterprise pays fees
2) How to build an explorer URL so a user can look at the on chain
balance of their Enterprise fee address

This commit adds the ENTEPRISE_PAYS_FEES coin feature, and adds this
feature to XTZ and ETH (the only two coins which currently pay fees from
the Enterprise gas tank). This solves for case #1.

This commit also adds a new optional field to `AccountNetwork` called
`accountExplorerUrl`. This can be used for building urls just like the
`explorerUrl`, and therefore solves case #2 above.

Ticket: BG-19107