Add missing fields in Eth Transaction
There are some fields in the base transaction class which were not set
in the Eth Transaction class. Most importantly, the _id field which
allows users of the library to get the transaction id of the
transaction. This commit implements a function in EthTransaction which
updates those fields in Eth Transaction

Ticket: BG-21259