Improve transaction builder speed
The ETH transaction builder was very slow for wallet initialization
transactions because we were parsing data (which can be huge) using
BigNumber, which is slower the bigger the number gets. In fact, we could
keep the data parsing logic very simple here as we just need the hex
string.

Ticket: BG-21437