Fix statics yarn build
Statics build was failing because of a flag that was being always
overridden by the spread operator because the object being spread always
had the same flag set (by definition). None of this code changed
recently, so I'm assuming this issue was caused by some update to
typescript or something.