Implement sign and explain for ETH-like coins
This commit adds an implementation of signTransaction and
explainTransaction for ETH-like coins relying heavily on logic in
bitgo-account-lib. The transaction explanation logic in
bitgo-account-lib is not fully finished yet, but that doesn't block an
initial implementation here -- when that is updated, we can simply add
a new change with the new verion of the library and extend our tests to
cover the newly explained fields.

This commit also bumps the version of dependency @bitgo/account-lib to
include new fixes required for signing transactions

Ticket: BG-22079