Add CGLD Celo token
The CGLD coin on CELO chain is both the native asset of the chain AND
a token following the ERC20 interface standard. BitGo intends to *send*
the coin using the ERC20 token interface, but must support receives of
it from both the token interface and the native asset interface. This
commit adds the token version of CELO to @bitgo/statics, including its
token address on celo mainnet chain

Ticket: BG-22179