Add token contract address to sends for CGLD
The transferBuilder accepts a coin name and susses out the token
contract idea from that using bitgo/statics if the coin is a token.
However, CGLD is *not* just a token in statics, but also a native
account coin. So this commit changes the builder function to be more
generic and accept token contract address for any type of coin that has
it.

Ticket: BG-22200