Make CELO not a "token"
CELO is a weird edge case in that it is both the native coin on its
chain and follows the ERC20 token interface. Thus we had to make a new
class for it to illustrate that. However, the coin should not be
specified as a "token" as that causes some components to fail, as they
assume that all "token"s as children of another coin and thus should not
be displayed as first class citizens. This commit changes CELO to not be
considered a token in this way

Ticket: BG-22435