Fix crypto deprecation warnings caused by __importStar
Using `import * as crypto from 'crypto';` imports all fields from the
'crypto' library into the current scope. This includes some deprecated
fields which emit a DeprecationWarning to stdout when the field is
imported, including when using import star. This is discoverable when
running node with the --trace-warnings flag when these warnings are
emitted.

Instead, just import the field desired. This should prevent the
deprecated field from being imported.

Ticket: BG-22502