Add support for flush tokens transactions
This commit adds support for flushing tokens on a wallet contract. This
is a function called flushForwarderTokens which allows the wallet to
pull in any tokens that are stuck on forwarder contracts. The builder
must accept a custom token address as we often need to flush
unsupported tokens (i.e. @bitgo/statics will not know about them).

Ticket: BG-22677