Fix ETH Clones signatures
This commit fixes ETH clones signatures when parsing from a serialized
transaction. Transfer builder has subclasses which include the
coinspecific signature salt to avoid replay protections. In this
particular case when building from serialized, we were always
instantiating the ETH builder not the coin-specific ones, so it was
using the ETH salt instead of the coin specific salt. This commit fixes
that by using the `transfer` builder function always to instantiate
transfer builder and never instantiating them direction from the
construcotr

Ticket: BG-22729