Fix CELO receive address generation
When CELO addresses are generated, it throws a validation error but
still actually generates the address. This is because it tries to
"verify" the address, but since it is pending chain initialization the
address is not there yet. This commit changes the createAddress function
to skip verification for addresses which are pending chain
initialization

Ticket: BG-22647