Remove webhook in wallet integration test
There is a wallet integration test that creates a new webhook and checks
that it works. However, it is stateful and if anything happens in the
middle of a test, for example it gets canceled - it's possible that the
webhook will be left in existence and all future tests will fail due to
this. To avoid this issue, this commit adds a before hook which removes
the webhook, hopefully making it self-healing

Ticket: BG-22987