Update @celo/contractKit for accountlib
This commit updates the @celo/contractKit library for accountlib to its
newest version, and includes the fixes for API changes that have
happened since then. This is because there was a bug in the previous
version that caused some perfectly valid signatures to be considered
invalid, which threw and caused user transactions to fail and flaky
tests in other systems. This commit also includes one such transaction
here, which previously would have failed due to the above error.

Ticket: BG-23095