Add support for verifying custom change addresses
Custom change addresses are addresses derived from another wallet's
keys, but should still be considered an internal spend. In order for
this to be safe, the custom change xpubs must be signed prior to sending
the transaction, but for now we'll just trust that the custom change
wallet is what the user wants it to be.

This security issue is known and will be corrected in the next commit
by verifying the change address key signatures if any output address is
found to be derived from them.