Require valid custom change key signatures when necessary
If a change address was successfully derived using they custom change
keychains, the signatures for those keychains must be validated against
the spending wallet's user key to ensure they have not been tampered
with.

This commit enabled mandatory verification of the custom change key
signatures when the transaction spends to an address derived from the
custom change keychains.

Additionally, the check for a match between the claimed user xpub
and the decrypted user xprv is made mandatory in this case.