feat: Fix CELO token transactionBuilder
BitGoJS is currently nonconformant with our standard naming scheme of:
- getChain() = the base chain that the coin lives on
- getName() = the coin name

Instead, getChain() returns the token name for tokens and the chain name
for native coins. This was causing an issue instantiating accountlib
transactionbuilder objects which expected the base chain. Ideally the
solution would involve refactoring the naming scheme to conform to
standard, though this would require changes to ERC20 and XLM tokens as
well, and be a breaking change. This would likely cause integration
issues both for us and clients.

This commit fixes the issue in a non-breaking way by adding a new
function getBaseChain() for abstractEthLike coins and tokens which
returns the expected base chain. This is the value used for
instantiating accountlib transactionBuilders now and should fix the
issue

CLOSES TICKET: BG-25905