fix(express): remove gcompat, switch to alpine build container
gcompat is not meant to be a long term solution for providing glibc
compatibility for compiled binaries. Instead, we should compile these
binaries in an environment which has the same libc implementation as the
one which will be present in the final production container.

This change replaces the `node:10` build container with
`node:10-alpine`, removes `gcompat` from the production container.

Ticket: BG-26745