fix: replace sed with js function for replacing unsafe evals
The sed -i"" syntax for in-place replacement of unsafe eval code in the
webpack pipeline is not portable. This caused the unsafe eval code to
remain in the browser bundle since the release process was recently
switched onto an ubuntu machine.

Instead, replace invocation of sed -i"" with some simple js code that
is more portable.

Ticket: BG-27261