feat(express): add support for returning keychains with generated wallet
It's convenient for callers of `wallet/generate` to be able to get the
keychains for their newly generated wallet in the response, instead of
needing to make additional API calls.

To do this in a backwards-compatible way, this change adds a
`includeKeychains` query parameter which, if set, will return the return
of the SDK `generateWallet()` call directly, which includs the wallet
keychains.

Ticket: BG-28844