fix(account-lib): change statics version back to ^6.0.0
This version should match the version in modules/statics/package.json.

The version is bumped as part of the release process, so overriding it
in @bitgo/account-lib has no effect. The downside is that this causes
yarn to generate an undesired @bitgo/statics entry in the yarn.lock.

Ticket: BG-27462