Enable batching eth transfers
Wallet platform was denying sendMany calls because the isBatch field was not included in the input of signTransaction. The isBatch param is returns when calling /tx/build in wallet platform and is set automatically if the length of recipients is more than 1.

This commit fixed this issue by adding isBatch to the output of signTransaction and would pass the param value from tx/build.

Ticket: BG-27047