fix(express): always prefer command line arguments to env var args
When the "IPC" argument was added, all subsequent configs had their
index increased by one. This revealed a bug where odd-numbered config
item indexes incorrectly preferred the env var arguments to the command
line arguments. Because the env vars defaulted to testnet, this broke
the ability to start bitgo express in the production configuration with
the `-e` flag on the command line.

Instead of reversing the configs in the `get()` call, just provide the
configs in the correct order to start with.

Ticket: BG-29084