ci: fix ci scripts to work for PR events
These were making the assumption that the PR branch was not yet merged
to master, which isn't the case for PR builds (only push builds).

Please note that this isn't actually merging the PR into the master
branch as if the PR was accepted and merged - this is only a tentative
merge which exists ephemerally within drone during the build pipeline.

Ticket: BG-30341