fix(core): update yarn resolutions to temporarily resolve audit issues
The recently disclosed vulnerability in `netmask` is causing the `yarn
audit` build step to fail, blocking all changes to the SDK.

While we wait for upstream packages to push updates, update the yarn
resolutions to use `netmask@^2.0.1` which includes a fix for the
disclosed vulnerability.

Additionally, resolve elliptic to `elliptic@^6.5.4` to work around
similar (but lower severity) issue in that package.

Note that this is not a long term fix and doesn't resolve the underlying
vulnerability for users of SDK packages (they will get similar errors if
they run `yarn audit` or `npm audit` in their own packages, since the
yarn resolutions and yarn.lock won't be respected).

Ticket: BG-30717