fix(express): don't store false when boolean flags are not given
The config precedence system expects that command line options which
aren't present to not exist (ie, set to `null` or `undefined`), but when
we use the `storeTrue` action for argparse arguments, omitted options
get set to `false`, which causes them to override any values passed via
environment variables.

Instead, use `storeConst` and store `true` explicitly if the option is
given. This will cause argparse to set the argument value to `null` if
it is omitted.

Ticket: BG-30729