feat(eth): generalize chain id configuration
This commit generalizes the creation of the "Ethereum Common" object
which sets the chain id and other fields on transactions. Previously it
didn't support goerli testnet and was automatically using the kovan
configuration. This commit generalizes the creation of the common object
such that it can accept any chain type we give it, as defined in statics

CLOSES TICKET: BG-31300