chore: move `typescript` and `ts-node` deps to root level
These are used by pretty almost all the packages in the monorepo, and we
should not be using different typescript versions for different packages
as it can make things very confusing.

Instead, bring the `ts-node` and `typescript` dependencies up to the top
level so we can use a consistent version across the whole monorepo.

Ticket: BG-31710