feat(account-lib): Add latest version of algo-sdk
We want to replace all usages of the algosdk fork that we have with account-lib and have account-lib
consume the latest version of algosdk directly. This is because we don't want to maintain this fork
anymore and we want to standardize the implementation for all our coins to use account lib.

This change adds the latest version of algo-sdk to account lib. Imports in the core package were
updated to continue using the fork.

ticket: BG-31597