feat(account-lib): add support for algo flat fees
Algorand allows you to set fees in two different ways. One is a
flat fee where it is a flat fee for the whole transaction and
the other is a fee per byte in microalgos. Previously we
assumed it was always a flat fee.

This change updates the logic in transaction builder to account
 for these two different fee types.

ticket: BG-31556